Read on...
Every tick the jumper will move some distance. It might cross over onto
a Quad obstical. 

The methods I have been adding and removing are to aid in calcuating how
to fix these kinds of crosses before they happen. 

The strategy is to calculate where the next tick will move the jumper,
and see if it intersects any known obsticle.