Over the next few commits, I'm going to show my work. READ MORE HERE
If you want to follow along, I advise you to get out a piece of lined
graphing paper.

Draw a quadrant I on this paper, include point (5,5).

Draw a point at (1,4). This is the player position.
model.getPlayerXPos(), for example, will return 1

Draw a rectangle from coodrinate (0,0) to (4,1). This will be the
stickyQuad. getX2(), for example, will return 4


Now you are ready.