Step 01 - Relative position of stickyQuad
Each variable (x1,x2,y1,y2) is calculated to be the relative position of
that coordinate of the stickyQuad.

Running MainTester will print out useful information. See how y2 is set
to -3. This means that to get from the player location to the y2 edge of
the stickyQuad you have to go down (negative) on the y axis 3 units. 

Verify the other 3 answers are correct.