"Stationary mouse makes 4 frame jump" bug found
Description:

The animation in the game grinds to a near halt. Jumping straight up
results in 4 (4!) distinct frames of movement. No discepency is found in
update times.


Extra odd:

The bug only occures when the mouse is stationary. Moving the mouse
while the square is moving causes the movement to become fluid.