Bug identified
Blocks would not appear when I ran the program.

I changed MainTester to only only run GamePanel

Blocks still did not appear. 

Printing out things revealed the height (calclulated by y2-y1) was zero
for ALL blocks. 

This is highly suspicious. 

A printout in the constructor of Quad (superclass for Block, so it will
be called for ever block created) showed that y2==y1 the second time a
bunch of Quads are created.

Throwing a strategic error let's me tell the exact stack trace when the
error was introduced.

Next commit has fix and description of fix.