Use null to represent no movement necessary
I tried to use Point.ZERO, but ran into a case where the movement would
be zero. 

Filtering zeros from my list of possible movements broke the mecahnic.

This results in a bug. If you click towards a wall the jumper is already
stuck on, the game doesn't record that as a stick. The player freezes in
fall. Weird bug.



I could Optional to represent this idea, but I concously choose not to.

The purity of Optional representing something that could be there is
secondary to the impracticality of having a return type of
Optional<Optional<Point>>. 

Java is verbose, but not that ugly. 

The best solution would be if Optional had a Optional.reduce(T a)
method, that constructs a single Optional out of nested optionals. 

Similar to >>= in haskell